requirements.txt
fastapi
uvicorn
langgraph
langchain-google-genai
python-dotenv
pydantic
pydantic-core
langchain-groq
langchain-mistralai
qdrant-client
langchain-qdrant